PythonApi/requirements.txt
flask
mastodon.py
beautifulsoup4
scikit-learn
python-dotenv
tenacity
iointel
